FIRST AID FINAL PROJECT/requirements.txt
langchain_core
langchain_groq
flask
langchain
python_dotenv
